+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Unreleased Changes
+------------------
+
+* Document release process
+
 0.2.0 (2017-11-30)
 ------------------