+++ CHANGES.rst
@@ -5,6 +5,7 @@ Unreleased Changes
 ------------------
 
 * Document release process
+* Document how to run in ECS as a Scheduled Task
 
 0.2.0 (2017-11-30)
 ------------------