+++ CHANGES.rst
@@ -6,7 +6,8 @@ Unreleased Changes
 
 * Document release process
 * Document how to run in ECS as a Scheduled Task
-* ``LocalCommand`` class - if ``script_source`` parameter is specified, instead of ignoring ``command`` send it as arguments to the downloaded script.
+* ``LocalCommand`` - if ``script_source`` parameter is specified, instead of ignoring ``command`` send it as arguments to the downloaded script.
+* ``LocalCommand`` bugfix - Handle when retrieved script_source is bytes instead of string.
 
 0.2.0 (2017-11-30)
 ------------------