+++ CHANGES.rst
@@ -8,6 +8,7 @@ Unreleased Changes
 * Document how to run in ECS as a Scheduled Task
 * ``LocalCommand`` - if ``script_source`` parameter is specified, instead of ignoring ``command`` send it as arguments to the downloaded script.
 * ``LocalCommand`` bugfix - Handle when retrieved script_source is bytes instead of string.
+* Add ``-m`` / ``--only-email-if-problems`` command line argument to allow suppressing email reports if all jobs ran successfully.
 
 0.2.0 (2017-11-30)
 ------------------